move alias to separate file
